Made AccessTokenResponse type to parse response. Need to store values for user now.
