+++ spoti-friends/src/auth/utils/AuthErrors.swift
